+++ CHANGELOG.rst
@@ -28,7 +28,7 @@ Fixed
 -----
 - added timeout to terminal input channel to avoid freezing input in case of server
   errors
-- Fill slots in other parts of utterance templates than text
+- Fill slots for image template other than text
 
 [1.1.7] - 2019-07-18
 ^^^^^^^^^^^^^^^^^^^^